voice: extract AppShell from ChatScreen
Pulls the outer layout (header, conversation sidebar, card container)
into a reusable shell so voice can adopt the same chrome next.
ChatScreen is now a thin wrapper that renders its message area,
rate-limit indicator, and ChatInput inside AppShell.

ConversationPanelState moves to AppShell as the canonical home;
use-conversation-panel-state updated to import from there.
ChatScreen rendering is pixel-identical — existing ChatScreen tests
still cover the header/sidebar branches via the shell.
