+++ webui/src/components/chat/ChatScreen.tsx
@@ -1,23 +1,22 @@
 // Producer Pal
 // Copyright (C) 2026 Adam Murray
+// AI assistance: Claude (Anthropic)
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 import { useEffect, useState } from "preact/hooks";
 import {
+  AppShell,
+  type ConversationPanelState,
+} from "#webui/components/AppShell";
+import {
   type MessageOverrides,
   type RateLimitState,
 } from "#webui/hooks/chat/use-chat-types";
-import { useUpdateCheck } from "#webui/hooks/use-update-check";
 import { type UIMessage } from "#webui/types/messages";
 import { ChatStart } from "./ChatStart";
-import { ChatHeader } from "./controls/ChatHeader";
 import { ChatInput } from "./controls/ChatInput";
 import { type HeaderInfo } from "./controls/header/HeaderActions";
 import { RateLimitIndicator } from "./controls/RateLimitIndicator";
-import {
-  ConversationPanel,
-  type ConversationPanelProps,
-} from "./ConversationPanel";
 import { MessageList } from "./MessageList";
 
 /**
@@ -43,16 +42,6 @@ interface ChatScreenProps {
   showTimestamps: boolean;
   showTokenUsage: boolean;
   conversationPanel: ConversationPanelState;
-}
-
-/** State and handlers for the conversation history panel.
- * Extends ConversationPanelProps with renamed callbacks and panel toggle state. */
-export interface ConversationPanelState extends Omit<
-  ConversationPanelProps,
-  "onNewConversation"
-> {
-  onToggle: () => void;
-  onNew: () => void;
 }
 
 /**
@@ -98,106 +87,61 @@ export function ChatScreen(props: ChatScreenProps) {
     showTokenUsage,
     conversationPanel,
   } = props;
-  const latestVersion = useUpdateCheck();
   const [thinking, setThinking] = useThinkingOverride(props);
 
   const currentOverrides: MessageOverrides = {
     thinking,
   };
 
-  const activeConv = conversationPanel.activeConversationId
-    ? conversationPanel.conversations.find(
-        (c) => c.id === conversationPanel.activeConversationId,
-      )
-    : undefined;
-
   return (
-    <div className="flex flex-col h-screen bg-zinc-100 dark:bg-zinc-950">
-      <ChatHeader
-        headerInfo={headerInfo}
-        mcpStatus={mcpStatus}
-        isHistoryOpen={conversationPanel.isOpen}
-        isActiveBookmarked={activeConv?.bookmarked}
-        latestVersion={latestVersion}
-        onOpenSettings={onOpenSettings}
-        onOpenToolsSettings={onOpenToolsSettings}
-        onOpenConnectionSettings={onOpenConnectionSettings}
-        onToggleHistory={conversationPanel.onToggle}
-        onNewConversation={conversationPanel.onNew}
-        onToggleBookmark={
-          conversationPanel.activeConversationId
-            ? () =>
-                conversationPanel.onToggleBookmark(
-                  conversationPanel.activeConversationId as string,
-                )
-            : undefined
-        }
-      />
-
-      <div className="flex flex-1 min-h-0 justify-center">
-        <ConversationPanel
-          isOpen={conversationPanel.isOpen}
-          conversations={conversationPanel.conversations}
-          activeConversationId={conversationPanel.activeConversationId}
-          onSelect={conversationPanel.onSelect}
-          onNewConversation={conversationPanel.onNew}
-          onDelete={conversationPanel.onDelete}
-          onExportItem={conversationPanel.onExportItem}
-          onRename={conversationPanel.onRename}
-          onToggleBookmark={conversationPanel.onToggleBookmark}
-          onExport={conversationPanel.onExport}
-          onImport={conversationPanel.onImport}
-          notification={conversationPanel.notification}
-          onDismissNotification={conversationPanel.onDismissNotification}
-        />
-
-        <div
-          className={`flex flex-col flex-9999 min-w-0 max-w-5xl ${conversationPanel.isOpen ? "hidden sm:flex" : ""}`}
-        >
-          <div className="flex flex-col flex-1 min-h-0 w-full border-x border-zinc-300 dark:border-zinc-700 shadow-lg dark:shadow-black/40 bg-white dark:bg-zinc-900">
-            <div className="flex-1 overflow-y-auto">
-              {messages.length === 0 ? (
-                <ChatStart
-                  mcpStatus={mcpStatus}
-                  mcpError={mcpError}
-                  checkMcpConnection={checkMcpConnection}
-                  handleSend={handleSend}
-                  overrides={currentOverrides}
-                />
-              ) : (
-                <MessageList
-                  messages={messages}
-                  isAssistantResponding={isAssistantResponding}
-                  handleRetry={handleRetry}
-                  handleEdit={handleEdit}
-                  showTimestamps={showTimestamps}
-                  showTokenUsage={showTokenUsage}
-                  requestedModel={headerInfo.activeModel}
-                />
-              )}
-            </div>
-
-            {rateLimitState?.isRetrying && (
-              <RateLimitIndicator
-                retryAttempt={rateLimitState.attempt}
-                maxAttempts={rateLimitState.maxAttempts}
-                retryDelayMs={rateLimitState.delayMs}
-                onCancel={onStop}
-              />
-            )}
-
-            <ChatInput
-              handleSend={handleSend}
-              isAssistantResponding={isAssistantResponding}
-              hasError={conversationHasError(messages)}
-              onStop={onStop}
-              thinking={thinking}
-              onThinkingChange={setThinking}
-            />
-          </div>
-        </div>
+    <AppShell
+      headerInfo={headerInfo}
+      mcpStatus={mcpStatus}
+      conversationPanel={conversationPanel}
+      onOpenSettings={onOpenSettings}
+      onOpenToolsSettings={onOpenToolsSettings}
+      onOpenConnectionSettings={onOpenConnectionSettings}
+    >
+      <div className="flex-1 overflow-y-auto">
+        {messages.length === 0 ? (
+          <ChatStart
+            mcpStatus={mcpStatus}
+            mcpError={mcpError}
+            checkMcpConnection={checkMcpConnection}
+            handleSend={handleSend}
+            overrides={currentOverrides}
+          />
+        ) : (
+          <MessageList
+            messages={messages}
+            isAssistantResponding={isAssistantResponding}
+            handleRetry={handleRetry}
+            handleEdit={handleEdit}
+            showTimestamps={showTimestamps}
+            showTokenUsage={showTokenUsage}
+            requestedModel={headerInfo.activeModel}
+          />
+        )}
       </div>
-    </div>
+
+      {rateLimitState?.isRetrying && (
+        <RateLimitIndicator
+          retryAttempt={rateLimitState.attempt}
+          maxAttempts={rateLimitState.maxAttempts}
+          retryDelayMs={rateLimitState.delayMs}
+          onCancel={onStop}
+        />
+      )}
+
+      <ChatInput
+        handleSend={handleSend}
+        isAssistantResponding={isAssistantResponding}
+        hasError={conversationHasError(messages)}
+        onStop={onStop}
+        thinking={thinking}
+        onThinkingChange={setThinking}
+      />
+    </AppShell>
   );
 }
