+++ webui/src/hooks/chat/helpers/use-conversation-panel-state.ts
@@ -4,7 +4,7 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 import { useMemo } from "preact/hooks";
-import { type ConversationPanelState } from "#webui/components/chat/ChatScreen";
+import { type ConversationPanelState } from "#webui/components/AppShell";
 import { type UseConversationsReturn } from "#webui/hooks/chat/use-conversations";
 import { type ViewState } from "#webui/hooks/use-view-state";
 import { isMobile } from "#webui/utils/is-mobile";
